Added separate header file for class vertex.
Vertex was before a private class defined inside the graph class.
The problem was that to insert a vertex into the graph
the DAO (database access object) would need to return an array
of objects or call the DAO multiple times to get all the data.

To solve the problem, the vertex class was separated from the
graph class. This way the DAO can return a vertex and then add it
to the graph.