Main is now capable of returning multiple outputs depending on the option passed.
Before the main always returned the svg representation of the best solution.
Now, depending on the option that is passed by the user, the main is capable
of returning the best solution, the best cost, the whole process of the
heuristic to find the best solution, and the svg representation of the best
solution.